+++ example/lib/main.dart
@@ -19,9 +19,20 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
-    controller = CanvasController((pngBytes) {
-      print('PNG bytes exporting canvas: $pngBytes');
-    });
+    controller = CanvasController(
+      (pngBytes) {
+        print('PNG bytes exporting canvas: $pngBytes');
+      },
+      onUndo: () {
+        print('Undo action triggered - validation flags reset');
+      },
+      onRedo: () {
+        print('Redo action triggered - validation flags reset');
+      },
+      onErase: () {
+        print('Erase mode enabled - validation flags reset');
+      },
+    );
   }
 
   @override