+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import flutter_advanced_canvas_editor
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FlutterAdvancedCanvasEditorPlugin.register(with: registry.registrar(forPlugin: "FlutterAdvancedCanvasEditorPlugin"))
+}