+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,14 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <flutter_advanced_canvas_editor/flutter_advanced_canvas_editor_plugin_c_api.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  FlutterAdvancedCanvasEditorPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FlutterAdvancedCanvasEditorPluginCApi"));
+}