+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## 1.0.6
+## 1.1.6
 
 CHANGELOG:
-- adjustment of the dots size
-- scale delete/rotate icons
+- Undo Calls onUndo?.call() when undo is triggered
+- Redo Calls onRedo?.call() when redo is triggered
+- Erase Calls onErase?.call() when erasing mode is enabled
 
 * TODO: export in different formats and colors